requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
psutil>=5.9.0
tqdm>=4.66.0 